% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{check_link_exists}
\alias{check_link_exists}
\title{Check that a link exists}
\usage{
check_link_exists(rd, xpath_sel, thing_type = "", quiet = F)
}
\arguments{
\item{rd}{a remoteDriver object}

\item{xpath_sel}{xpath selector for the object}

\item{thing_type}{character describing the link object, e.g. "bullet" - used in message passing}

\item{quiet}{suppress messaging}
}
\description{
Returns the link url if the link exists and NA otherwise
}
