% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{create_bullets}
\alias{create_bullets}
\title{Function to create a new bullet in a specific barrel}
\usage{
create_bullets(rd, df)
}
\arguments{
\item{rd}{a remote driver}

\item{df}{a single-row data frame with the necessary information (bullet, cartridge_des,
lot_no, firing_seq, comments, brand, brand_other, caliber, caliber_other, grain,
surface_mat, surface_mat_other, barrel, details_url)}
}
\description{
Function to create a new bullet in a specific barrel
}
