% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{create_firearms}
\alias{create_firearms}
\title{Function to create a new firearm}
\usage{
create_firearms(rd, df, seturl = rd$getCurrentUrl())
}
\arguments{
\item{rd}{a remote driver}

\item{df}{a single-row data frame with the necessary information (name, model,
comments, brand, brand_other, caliber, caliber_other, consec_manufacture,
cartridges, bullets, breech_face_class, breech_face_other, firing_pin_class,
firing_pin_other, n_lands, twist_direction)}

\item{seturl}{URL to go to before creating the barrel (defaults to current URL)}
}
\description{
Function to create a new firearm
}
