% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{fill_checkbox}
\alias{fill_checkbox}
\title{fill in a checkbox}
\usage{
fill_checkbox(rd, css_sel, value)
}
\arguments{
\item{rd}{a remoteDriver object}

\item{css_sel}{css selector for the entry box}

\item{value}{value to fill in}
}
\description{
If value is true, check the box, otherwise, nothing
}
