% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{fill_conditional_field}
\alias{fill_conditional_field}
\title{Fill in a dropdown field that has an "other" option text box}
\usage{
fill_conditional_field(rd, css_sel_init, css_sel_other, value_init,
  value_other)
}
\arguments{
\item{rd}{a remoteDriver object}

\item{css_sel_init}{css selector for the dropdown box}

\item{css_sel_other}{css selector for the "other" text box entry}

\item{value_init}{value for the dropdown box}

\item{value_other}{value for the text box (NA if dropdown value doesn't trigger the text box)}
}
\description{
Fill in a dropdown field that has an "other" option text box
}
