% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{fill_dropdown_field}
\alias{fill_dropdown_field}
\title{fill in a dropdown field}
\usage{
fill_dropdown_field(rd, css_sel, value)
}
\arguments{
\item{rd}{a remoteDriver object}

\item{css_sel}{css selector for the entry box}

\item{value}{value to fill in}
}
\description{
Function will check to make sure that the value specified is available (matches the beginning of an option)
and will issue a warning if this is not the case.
}
