% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{fill_logical_field}
\alias{fill_logical_field}
\title{fill in a logical dropdown field}
\usage{
fill_logical_field(rd, css_sel, value)
}
\arguments{
\item{rd}{a remoteDriver object}

\item{css_sel}{css selector for the entry box}

\item{value}{value to fill in}
}
\description{
Sends only the first character of value (T/F)
}
