% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{fill_text_field}
\alias{fill_text_field}
\title{fill in a text field}
\usage{
fill_text_field(rd, css_sel, value, clear = F)
}
\arguments{
\item{rd}{a remoteDriver object}

\item{css_sel}{css selector for the text entry box}

\item{value}{value to fill in (character)}

\item{clear}{should the text box be cleared before entering the new value?}
}
\description{
fill in a text field
}
