% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{find_text}
\alias{find_text}
\title{Get text for a link}
\usage{
find_text(rd, css_sel)
}
\arguments{
\item{rd}{a remote driver}

\item{css_sel}{CSS selector for link element}
}
\description{
Get text for a link
}
