% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{nbtrd_login}
\alias{nbtrd_login}
\title{Log in to NBTRD page}
\usage{
nbtrd_login(rd, user, password)
}
\arguments{
\item{rd}{a remoteDriver object that is at the NBTRD homepage}

\item{user}{username}

\item{password}{password}
}
\description{
Log in to NBTRD page
}
