% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{open_modal}
\alias{open_modal}
\title{A loop test to open a modal for data upload}
\usage{
open_modal(rd, modal_css_sel, btn_css_sel)
}
\arguments{
\item{rd}{a remote driver}

\item{modal_css_sel}{css selector for the modal (usually ".modal-open")}

\item{btn_css_sel}{css selector for the button to open the modal}
}
\description{
A loop test to open a modal for data upload
}
