% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{parse_objective}
\alias{parse_objective}
\title{Parse a string containing magnification value 'dddX' into a number 'ddd'}
\usage{
parse_objective(x)
}
\arguments{
\item{x}{string}
}
\value{
number extracted from x
}
\description{
Parse a string containing magnification value 'dddX' into a number 'ddd'
}
