% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{quiet_try}
\alias{quiet_try}
\title{Suppress all messages and errors when running an expression}
\usage{
quiet_try(expr)
}
\arguments{
\item{expr}{Expression}
}
\description{
Suppress all messages and errors when running an expression
}
