% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_helper.R
\name{setup_NBTRD}
\alias{setup_NBTRD}
\title{Set up connection to NBTRD}
\usage{
setup_NBTRD(selenium_port = 4445L, strict = T)
}
\arguments{
\item{selenium_port}{integer port number of docker container}

\item{strict}{error if RSelenium connection does not succeed?}
}
\value{
RSelenium remoteDriver object, if connection was successful, NULL otherwise.
}
\description{
This function connects to a pre-existing Docker firefox container running
on localhost:4445, and if none is found, it uses the RSelenium client option,
(which is deprecated) to connect.
}
