% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{submit_modal}
\alias{submit_modal}
\title{A loop test to close a modal and submit the data}
\usage{
submit_modal(rd, modal_css_sel, btn_css_sel)
}
\arguments{
\item{rd}{a remote driver}

\item{modal_css_sel}{css selector for the modal (usually ".modal-open")}

\item{btn_css_sel}{css selector for the button to close the modal/save the
data}
}
\description{
A loop test to close a modal and submit the data
}
